+++ PS.MTA-STS/functions/New-PSMTASTSFunctionAppDeployment.ps1
@@ -60,8 +60,8 @@
 
         [Parameter(Mandatory = $true)]
         [ValidateScript({
-                if ($_.length -lt 1 -or $_.length -gt 90 -or $_ -notmatch "^[a-zA-Z0-9\.-]*$") {
-                    throw "ResourceGroup name '$_' is not valid. The name must be between 1 and 90 characters long and can contain only letters, numbers, hyphens and periods."
+                if ($_.length -lt 1 -or $_.length -gt 90 -or $_ -notmatch "^[a-zA-Z0-9-_.]*$") {
+                    throw "ResourceGroup name '$_' is not valid. The name must be between 1 and 90 characters long and can contain only letters, numbers, hyphens, underscores and periods."
                 }
                 else {
                     $true